fix: carpeta de colapsables
