+++ .storybook/preview.ts
@@ -24,7 +24,9 @@ const headers = [
   'componentes-calendario--calendar-no-references',
   'componentes-calendario-con-eventos',
   'componentes-calendario',
-  'componentes-colapsable',
+  'componentes-colapsable-tipos',
+  'componentes-colapsable-encabezado',
+  'componentes-colapsable-contenido',
   'componentes-desplegables-de-navegación',
   'componentes-desplegables-de-selección',
   'componentes-destacado-banner',
