+++ src/scss/components/_cards.scss
@@ -790,7 +790,8 @@ button.card {
         }
 
         i,
-        .material-icons-round {
+        .material-icons-round, 
+        .collapse-img-icon {
           margin: auto 0;
           font-size: 1.5rem;
         }
